Added first full minigame + fact (Jurassic)
Worked on Button Class to add an image button (not integrated yet but soon), fully created jurassicEra with 1 minigame (FossilFinder) and fact, will add probably one more minigame and fact later if have we have time, but will do the robot one next. The buttons are just boxes right now, designs are in the process for buttons, will replace later for aesthetics - working on functionality for now. DinoDodger in process of being built.